Initial stages of repositioning ability
I remembered a joke today.  Its a historical joke, so people who like that sort of thing will love it.  Its also a futuristic sci-fi joke, so people who like that will love it.  Its also a funny joke, so people who like that sort of thing will love it.